ath11k: Print driver stats before system crash
Most of the times, having essential driver stats on firmware
crash would be useful for easier debugging purposes.
This patch adds this support by making use of
a ssr notifier which gets called on rproc crash.
Note that panic() which was called by the rproc_crash
handler will not happen in this case and ath11k driver
will induce a crash.
Additionally this patch also adds support for dumping
ring backpressure stats on crash.

Change-Id: If8f5c6f7b05c449e7243aa6be98a23831a7f795b
Signed-off-by: Sriram R <srirrama@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>