ath11k: poll reo status ring for IPQ5018
Currently reo status interrupts are not received
due to wrong mapping of the reo status interrupt
line in IPQ5018.

Hence, until the mapping is resolved in HW, use
polling to reap the reo status ring. Rather than
a period timer to reap the ring, the timer is
triggered only on sending a reo command with
status request.

Without proper reaping of the ring, backpressure
and ring full issues are seen in multi client test
setups which leads to flooding the console with
error messages reporting failure to send reo cmds.

Can be reverted once HW solution is available.

Change-Id: I5a8425a45abb3a4d45e3ba866c7e547e4764d566
Signed-off-by: Sriram R <srirrama@qti.qualcomm.com>>
Signed-off-by: Kaviarasu Sanjeevi <kaviaras@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>