ath11k: Fix unmapping allocated memory
In the present code we were not unmapping the allocated
memory, so in a scenerio like where we do infinite number
of FW crash and recovery, we face a crash as we run out of
memory and not able to allocate memory for futher mapping.
So, here we are cleaning up the old mappings while recovery is
happening.

Change-Id: I4a14adf1b460d2f22be177ffdd5a2ba0e914baa1
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>