ath11k: Fix rssi and retry count in station dump and skip tx bitrate update for mgmt frames
Fix signal and signal average in station dump by using real
noise floor and appropriate bandwidth offset. In this way
signal strength in station matches the rssi_in_dbm in htt stats
Tx bitrate in station dump was dropping to legacy rate
randomly, this is because the tx rate was updated for
management packet as well. Skip tx bitrate update for
management packets based on the tid.
Tx retries in station dump is showing 0 always.
Fixed it by getting the retry count in tx completion path
from tx descriptors. And updated this to station info in
sta_statistics mac ops.

Change-Id: Ia164793b92e247afa52b2f33e2b7da7c671cc5bc
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>