ath12k: Add support for mlo in cnss diag debug
Currently, when mlo is enabled when we run cnss_diag -p -c
radio index is always 0 (R0: FWMSG:...) due to this it is
not possible to identify radio and debugging is not possible.

Add ATH12K_TM_ATTR_LINK_IDX attribute to send ar->link_idx that is
fetched by cnss diag to display correct radio number.

Change-Id: If35abed2d5fba45638515cab541ae40f44003aef
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>