ath11k: add support to get rssi value of non-associated client
Monitor the surrounding network to get RSSI of non-associated
client(for example, if Neighbor-Peer associated to neighboring AP's).
Obtaining the RSSI it is useful for roaming between AP's in a
network and on deciding which AP the client should associate
with.
This is achieved through debugfs by providing the mac address
of a npr(neighbor peer) which creates a neighbor-peer object to store's the
data associated with the npr such as its mac, RSSI and timestamp.
Once the npr is added a file with the mac-address of npr as its
filename is created.
Each time the file is read a peer object is created for the
NPR and filter is set to get its ppdu_info for home channel
thereby obtaining the RSSI and a corresponding unix
timestamp is created, once the data collection is completed the
peer object is destroyed to free-up peer count.
Upto 6 NPR's can be created to keep the load and peer-count low.
Once the rssi and timestamp are obtained the peer object
created for NPR is deleted to free the peer id.
In case the NPR packets are unavailable during scan the
previous timestamped value is returned.

Command to add a Neighbor-Peer:
echo "add,<mac-addr-of-npr>" >
/sys/kernel/debug/ieee80211/phyX/ath11k/neighbor_peer

Command to delete a Neighbor-Peer:
echo "del,<mac-addr-of-npr>" >
/sys/kernel/debug/ieee80211/phyX/ath11k/neighbor_peer

Command to display RSSI of Neighbor-Peer:
cat /sys/kernel/debug/ieee80211/phyX/ath11k/nrp_rssi/<mac-addr-of-npr>

Change-Id: Iaefdde802415980a5d63dc715e103e7b6039c7f8
Signed-off-by: saleemuddin shaik <quic_saleemud@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>