ath12k: Add support to identify dual mac
While debugging using cnss diag for dual mac device, we are using Data.msc
file to get the logs. Instead we should be using Data_dual.msc file for dual
mac devices to get the logs.

Add a check if the device is dual mac and send it through nl command to be
sure that correct .msc file is used.

Change-Id: I49506a7d8b1faf7a3b6832f77519d011333b56ae
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>