ath11k : flush management frames to firmware before waiting for tx completion
warning print "ath11k c000000.wifi: failed to flush mgmt transmit queue 0"
is observed during interface down.

The management packets are queued in a skb_queue and the skb_queue
is dequeued in the work ar->wmi_mgmt_tx_work.

In ath11k_mac_flush_tx_complete, before waiting for the tx completion of
all the management frames, we are not ensuring that queued
management frames are flushed to the firmware.

This causes ar->num_pending_mgmt_tx to be positive and it leads to the
warning print.

Fix this by flushing all the management frames to firmware before waiting
for the tx completion.

Change-Id: Ia78d7a7057c27d5639e36f86e4a6defcd8361d5f
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>