ath11k: reap CE entries if ring is full and intrp idle for 500ms
If there is no buffer to send packet via HTC, then check if
interrupts are not processed from that CE for last 500ms.
If so, reap the CE ring and check if we could proceed for new enqueue.

This change is added because there is a case where intermittently
completion interrupts are not received from CE3 and hence adding
this WAR in host to come out of this issue scenario.

Change-Id: I46c8f79015d684128db9e155548a8cc7e49a7745
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>