ath12k: fix RU puncturing pattern for peers after CSA
If AP switches channel and with a valid puncturing pattern, the
peer at firmware level is not updated with the new value. This
leads to an issue where firmware still transmits the packets
in the whole bandwidth.

Fix the issue by updating the ru puncturing bitmap of all EHT
capable peers by sending Peer Channel width Switch command. The
command has puncturing member and hence it can be leveraged.

Change-Id: I93ae48f0e888fe4439e703555bce08e1f79aefdb
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>