ath11k: Add packet latency histogram support
MIME-Version: 1.0
Content-Type: text/plain; charset=UTF-8
Content-Transfer-Encoding: 8bit

The latency of a packet is measured from packet being enqueued
to the driver's transmit queue, to frame has been successfully
ack’ed by the receiver. Enqueue time is added in tx info and
difference in time is calculated once the packet is acked by
the receiver. Latency is calculated in powers of 2 and number
of packets that took a particular time is incremented in the
corresponding index and displayed via debugfs

Change-Id: I65ad835b529ff6addeb91739905b77fd6b5773d1
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>