wifi: ath11k: Add support to send the QoS Null Data frame through exception path
When we try to send QoS NULL Data frame in Ethernet encap type, it modified
as QoS Data frame with TID 0 (encryption enabled one if security enabled).
But expectation is, it should be send as open type frame with TID 7 since
its a QoS NULL data frame.

Added this frame under exception route to bypass TCL with the help of FW.

Change-Id: Iae4e655779c09f509404147d8056af2ee0dfc5f2
Signed-off-by: Karthikeyan Periyasamy <quic_periyasa@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>