wifi: ath12k: advertise support for SW progressive DFS
Advertise support to user space that the driver supports different
parameters for the device bandwidth compared to the operating bandwidth
and for background radar detection on the device bandwidth
when SW based progressive DFS is supported. Firmware will set
the service bit WMI_TLV_SERVICE_SW_PROG_DFS_SUPPORT which
determines if SW Progressive DFS is supported and
NL80211_EXT_FEATURE_DEVICE_BW in user-space will be updated.
Check for this support when background scan is requested, no need to
send WMI command to firmware as the CAC is requested on the current
device bandwidth where radar detection is already active.

Change-Id: If338c28dc48e709701f62cf77933609ceb5e59c1
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>