UPSTREAM:DP wifi: ath12k: Update DP IRQ names for
devices with hybrid bus.

In ath12k driver IRQ names for PCI is in the "pci%u_xxxx_%d" where
%u is based on the bus_id for PCI. For QCN6432 %u is based on
userpd_id.
In case of devices with both QCN9274 and  QCN6432,there is
a conflict as both the bus_id and userpd_id
is 1.Changed the IRQ name of QCN6432 from "pci%u_xxx"
to "pcic%u_xxx".

Earlier IPQ5332 and IPQ5424 was considered to rootPD, with
new implementation these are considered to first userPDs,
hence decreament userpd_id by 1 in certain places to maintain
parity with old implementation.

Patch dependency: 0122-UPSTREAM-DP-wifi-ath12k-Handle-irq-registration-for-hybrid-bus-t.patch
Change-Id: I139120e351a926ef0fbd39a35cf74473f646f9ea
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Narendhar Reddy <nare@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>