wifi: ath12k: Handle STA state machine in assign vif chanctx for STA mode
The STA state machine is handled for STA mode in the
assign vif chanctx ops in the driver.
This causes the firmware crash due the peer is deleted before
the peer assoc command in the unassign vif chanctx ops.
Hence, fix this issue by handling the STA state machine
for STA mode.

Change-Id: Ia394c9230382a61c6e21a7a7890c63b1b6c14f84
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>