wifi: ath12k: Configure channel dwell time during scan
Configure channel dwell time from duration of the scan request
received from mac80211 when the duration is non-zero. When the
scan request does not have a duration then use the max scan time
which is initialized to a default value of 20000ms.

Corresponding flag NL80211_EXT_FEATURE_SET_SCAN_DWELL is
advertised.

Change-Id: Ib1a056cde94aed21f323e8857c997046b5c8526a
Signed-off-by: Rohit Nair <quic_rohitnai@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>