mac80211: Fix for use-after-free in tx path
key memory is allocated and stored in two places,
  1. sta->ptk
  2. sta->fast_tx

After freeing the key memory, NULL pointer is assigned to sta->ptk but
not assigning to sta->fast_tx. Due to this, freed key memory accessed
in the tx path. To fix this issue, set tx key for old key after freed
the key memroy.

Change-Id: Ia045717472c78325ec7666c63b761dd973930cc8
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>