wifi: ath12k: Add Basic support for Bridge Peer
When ML peer associates on diagonal links, create bridge peer on the
device adjacent to each other. Bridge peer is an link peer created
on the device for every peer associating (which are not adjacent to
each other) which is known only to driver and firmware. Bridge peer
will be created on bridge vap only.

Summary of changes are as follows:
 1. Extend 2 more ath12k_link_sta links to accommodate bridge peer
    creation apart from existing 16 links (0-15).
 2. Add an logic to determine bridge peer is required or not for the
    every associating peer.
 3. Handle bridge peer creation, deletion and SSR recovery.
 4. Handle dp_tx and dp_rx specific to bridge peer.

When an ML peer associates on diagonal links in STA mode, create a bridge
peer on the adjacent device, similar to AP mode. Upon receiving the
Association response, the supplicant will request to add a link STA
followed by a SET STATION command.

The driver should check for adjacency to identify diagonally opposite links.
If such links are found, create a Bridge Peer by sending a WMI PEER Create
command with TYPE_DEFAULT. The firmware will send an HTT peer map event
for successful Bridge Peer creation.

Patch-dependency:
	0327-01-wifi-ath12k-n-link-mlo-bridge-vdev-basic-handling.patch
	0327-02-wifi-ath12k-n-link-mlo-bridge-vdev-create-start.patch
	0327-03-wifi-ath12k-n-link-mlo-bridge-vdev-up-and-channel-switch.patch

Tag: N_LINK_MLO

Change-Id: I70da486148898078b8fb8f2c9de31b1c986ca53c
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>