QSDK: wifi: ath12k: Add N-LINK-MLO WSI Congestion WAR
QCN9274 devices are connected for MLO communications using WSI (WLAN
Serial interface). This interface connects all the devices on the
board in a daisy chain and the interface is directional.

As an example, find the WSI traffic flow of 4 QCN9274 devices.
Note that direction of the WSI traffic flow may change depending on
the platform as well as the devices can be connected in any order.

		2 GHz                 5 GHz-lower
		  +----------←----------+
		  |                     |
		  | Example connection  |
		  ↓        of           ↑
		  |  4 QCN9274 devices  |
		  |                     |
		  |                     |
		  +---------→-----------+
		5 GHz-upper           6 GHz

Each device has an ingress WSI link coming from another device and
an egress WSI link going to another device. each device also
multiplexes (and demultiplexes) traffic coming from another device.

In MLD case, WSI congestion may happen when a device has WSI egress
traffic (where the current device acts as primary or bridge and is
the source of WSI traffic) and ingress traffic (where the current
device acts as bypass link for WSI traffic that will multiplex with
the egress traffic).When the congestion happens on the egress link,
the data is not reaped as fast enough from the ingress ring on the
device. This results in firmware crash.

Fix this issue by detecting congestion possibility where identifying
both bypass traffic received by the ingress WSI link and traffic from
current device passing through egress WSI link and limiting the MPDU
size.

  1. ath12k determines the number of egress traffic and ingress traffic
     for each device every time a peer is associated or disassociated
  2. ath12k reports to firmware by sending WMI command if there is any
     change in egress or ingress traffic counts of each device in WSI
     connected devices
  3. Firmware administers the MPDU limit.

Add support to calculate egress and ingress count for each device and
send wmi command

Details of the changes:
	1. Initialization and Deinitialization of WSI load information
	   database
        2. For each device, update the WSI order in the database at the
	   end of core start
	3. whenever a client is associated or disassociated,
		a) identify the primary link device and update egress
		   count for the primary device
		b) identify the secondary devices (non-primary links)
		c) For each secondary device, calculate the number of
		   hops from primary device and update ingress count
		   for all the hop devices
		d) Set notify as true for the count updated devices
	4. send Pdev WSI stats info WMI command to firmware for the
	   devices set with notify as true

Patch-dependency:
	0325-02-wifi-ath12k-Add-support-for-four-link-MLO-in-AP-MLD.patch
	0337-11-QSDK-wifi-ath12k-Add-WMI-support-to-notify-WSI-stats.patch

Patch-work: None

Change-Id: Ic100adbbed28c6a57be89951bb609fb14c8ad0c5
Signed-off-by: Mahendran P <quic_mahep@quicinc.com>
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>