wifi: ath12k: Increment num_rules by 2 only if weather channels are present
When a 5 GHz AP is configured with country Norfolk Islands (country
code NF), there is a warning trace observed.

In ath12k_reg_build_regd, if the DFS region is ETSI, num_rules is
increased by 2 to support the weather radar channels. However, in the
country code NF, no regulatory rule supports weather radar channels.
Consequently, is_valid_reg_rule returns false when the for-loop
iteration in is_valid_rd exceeds the actual number of regulatory rules
present in the rd. The start frequency and end frequency for a
regulatory rule beyond the last regulatory rule in the rd is 0.
This triggers a WARN_ON within __regulatory_set_wiphy_regd.

To address this issue, increment num_rules by 1 within the for-loop
in ath12k_reg_build_regd, but only when the DFS region is ETSI and
the regulatory rule supports weather radar channels. Re-allocate the
new_regd struct to accommodate the new rules added.

Tag: REGULATORY

Change-Id: Ib81c26f955803808c644497e4fd3ff1bddf88097
Signed-off-by: Amith A <quic_amitajit@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>