QSDK: wifi: ath12k: Send low ack event with master netdev index
In WDS, it is observed that the event low ack is sent with the
NL80211_CMD_NOTIFY_CQM command to userspace during SSR recovery.
The ifindex attribute sent along with the NL command is not
recognized by hostapd and is ignored.

This occurs because hostapd is unaware of the AP_VLAN interface created.
Since this event is reported for a connected WDS STA and its sdata is
mapped to the AP_VLAN sdata

To fix this, send the NL command with the Master AP’s ifindex in the
case of WDS.

Patch-dependency: none

Change-Id: Iae01841991b7608b7b9f4553f7d82585291de709
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>