wifi: ath12k: Add AHB coredump collection support
Snapshot of memory region during crash is required to debug the reason
for crash. Collect dumps of various segments define in DTS as dump-region.
Also collect the dump of userPD firmware region. Dump should be collected
in ELF 32bit format.

First physical, virtual address along with size is computed for various
segments and then ELF header is framed with this information. Dump is
uploaded to user space with help of devcoredump framework.

The dev_coredumpm has a callback registered, which would be invoked on
trying to copy the data. All the information including physical address,
virtual address and size are stored in array of struct. ELF header is
formed with this information and same is received in the callback.
Header is first copied and then various segments are dumped one after
the other. Segment to be copied is calculated based on the sizes of the
segments and total bytes already copied.

Change-Id: I9f97e2ef71782f441e648bee324d7d2709bf7f05
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>