UPSTREAM:CP:wifi:ath12k:Add status parsing in peer assoc conf ev
Firmware crash is seen due to ml_peer_id reuse. The sequence of events
is as follows:
1. Peer X associates to AP
2. X gets disassociated due to some reason and host sends peer delete
   WMI to the firmware
3. Firmware flags the peer X for deletion but doesnot complete the
   delete and doesnot reply back to host with peer unmap
4. When host timeouts due to peer unmap not received cleans the peer
   from driver hereby freeing the ml_peer_id allocated.
5. Current logic in driver is to allocate ml_peer_id which is free and
   when new peer Y associates, ml_peer_id which was just freed in #4.
6. When firmware receives peer assoc WMI with duplicate ml_peer_id
   firmware asserts.

As a stop gap solution, status field in peer assoc confirmation event is
used to send the status of peer assoc in such cases of ml_peer_id
duplication. Status will be sent as 0 in success case and 1 in failure
case. When failure is seen from firmware, peer assoc failure will be
reported to mac80211 subsequently denying the client assocation with
code 17 and sending peer delete to firmware when state changes from 1 to
0.

Patch-dependency: None
Patch-work: None

Change-Id: I7ce8f5e3d4a210c1bc2e6c6f8104da0d5b0192d6
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>