wifi:ath11k: Handle duplicate descriptor case in mon dest ring
When the software cookie from the MSDU link descriptor matches the
previous software cookie in the monitor destination ring, the physical
address for the corresponding ring is not stored. This results in invalid
memory access when a duplicate entry is detected in the monitor
destination ring. Fix it by caching the last physical address and
comparing it with the new physical address. Skip the monitor destination
ring processing if they are equal.

Patch-dependency: none

Change-Id: I708e0f08c271519135495e582908b877f69c25cf
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>