UPSTREAM:CP:wifi:ath12k: Add pdev_id param to FTM WMI to handle split PHY
In FTM(Factory Test Mode) mode, the SOC level pdev_id received instead of
the radio level pdev_id from the FTM segment header. The split PHY target
has PHYA0 and PHYA1 in a SOC. Therefore, we need to obtain a unique pdev_id
for the split PHY to send and receive data properly based on the pdev_id.

To fix this issue, add the pdev_id param in FTM WMI command based on the
service bit to handle the split PHY use cases.

Change-Id: I0d54f4738f50a33cfc8816d89ead1d2dd6087ef3
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>