wifi: ath12k: combine channel list for splitphy
When two Splitphy SoCs having below combination are combined into an MLO
group they will be part of same wiphy and hence the channel list
(wiphy->bands[]) will be common for all of the radios (ar).
1 - 2.4GHz/2x2 + 5GHz Low band 2x2
2 - 5GHz High band/2x2 + 6GHz/2x2

This brings in a situation where 5GHz band is split across two radio(ar)
and whenever a radio updates channel list it updated the wiphy->bands[]
in a way that only the channels that its HW are enabled.
This results in wiphy->bands[NL80211_BAND_5GHZ] always having channels
of last radio which updated the list.

Fix this by updating only the channels which belongs to its range whenever
a radio(ar) tries to update wiphy->bands.

Change-Id: I14b2f1ffa0ec747c1e6fab1df2ab1d598f728231
Signed-off-by: Rameshkumar Sundaram <rameshkumar.sundaram@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>