wifi: ath12k: Fix the handling of TX packets in Ethernet mode
When the virtual interface is configured in Ethernet mode, instruct
the firmware to use a different encapsulation mode for QoS null frames
and EAPOL frames received in native Wi-Fi format. For multicast or
broadcast packets in native Wi-Fi or raw format, set the MSDU
extension descriptor to override the bank register configuration
on a per-packet basis.

Change-Id: Ie44df36c74c3763b962a3732141bf3c29f79d642
Signed-off-by: Nithyanantham Paramasivam <quic_nithp@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>