QSDK:PLATFORM ath12k: Add support for collecting PHY logs along with QDSS
Introduce changes to the ath12k driver to enable
the collection of PHY logs along with QDSS. Update
the driver to handle additional arguments for PHY log
collection in the same QDSS sys debug file.

In case of split phy, two phy bins will be collected.
This bins are needed for the ADC validation.

Add new QMI structures and functions to manage QDSS trace
data requests and responses.

  HOST ----(QMI_WLANFW_QDSS_TRACE_OFF_V01)   --------> FW
  HOST <----(QMI_EVENT_QDSS_TRACE_REQ_DATA(QDSS)) ---- FW
  HOST <----(QMI_EVENT_QDSS_TRACE_REQ_DATA(PHY)) ----- FW
  HOST ----(QMI_WLFW_QDSS_TRACE_DATA_REQ_V01) -------> FW

Commands to get phy logs:
 QDSS:  echo 0 > /sys/kernel/debug/ath12k/<chip_name>/trace_qdss
 PHYA0: echo 0 0x40 > /sys/kernel/debug/ath12k/<chip_name>/trace_qdss
 PHYA1: echo 0 0x80 > /sys/kernel/debug/ath12k/<chip_name>/trace_qdss

Change-Id: Id29133e3d134a054494bbd73e3003fd1933aebb4
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>