wifi: ath12k: Check if ar is down before handling mac operations
Usually when the chip goes for a crash, upper layers might
not be aware of it. Hence, we need to handle scenarios when
the upper layers require some actions to be done and the chip
is down.

Check for FLUSH flags and then decide whether the ar context
has to be handled or not.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.3.1-00173-QCAHKSWPL_SILICONZ-1

Change-Id: I098ecee73bef9268cd9ac131171d77dd36622b20
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <maharaja.kennadyrajan@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>