wifi: mac80211: fix skb and sdata mismatch
Currently, skbs are getting tagged with wrong sdata, i.e even if the
sdata does not support a skb specific channel, we are tagging the sdata
along with that.

In issue scenario, 1 DUT was having 6G Mesh(root) + 2G AP and the other
DUT was having 2G Mesh(child), and the packets from 2G Mesh were seeing
in 6G Mesh, interface.

In this fix, tagging the sdata only if the respective skb's channel support
is present in the corresponding sdata.

Change-Id: Ifbad0bac5f4eb51426f6b6afb9b44bbc6bc2d44a
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <maharaja.kennadyrajan@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>