wifi: ath12k: Attach and detach ath12k_dp_link_peer to ath12k_dp_peer
As part of peer modularization in Unified Driver Framework, station view
for the data path is shown below:

    ath12k_dp_peer
	   |
	   |-> ath12k_dp_link_peer
	   |
	   |-> ath12k_dp_link_peer
	   |
	   |-> ath12k_dp_link_peer

This ath12k_dp_peer contains the fields used in the per pkt Tx Rx paths
applicable across all the links and maintains an array of
ath12k_dp_link_peer. Per pkt Tx and Rx path operates on ath12k_dp_peer.

ath12k_dp_link_peer contains the fields specific to a link and these are
mostly for telemetry and monitor usage.

This patch is to add support to attach and detach ath12k_dp_link peer to
ath12k_dp_peer as and when it is created and deleted.

Change-Id: I5068bb7f48b1f530fc23a86fc148290857261790
Signed-off-by: Harsh Kumar Bijlani <quic_hbijlani@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>