wifi: ath12k: Add callbacks in arch_ops for rx APIs
Add callbacks in arch_ops for rx APIs which are defined
in wifi7 and need to be called from the common code.
This is required, as these APIs will be part of
ath12k_wifi7.ko and the common module ath12k.ko needs
ops mechanism to call these APIs.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.3.1-00173-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Change-Id: I353ad67133cc3bb8ccfead6a3b60055dd009e024
Signed-off-by: Pavankumar Nandeshwar <quic_pnandesh@quicinc.com>
Signed-off-by: Ripan Deuri <quic_rdeuri@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>