wifi: ath12k: Add new infra for the rx path
The existing usage of rx_ops to reach the hal_apis at multiple places uses
function pointer redirections, which comes with performance penalty.

In order to avoid multiple function pointer indirection in performance critical
path, add an API to extract the required hal paramters in the rx path right
at the beginning and store them in a structure which is then leveraged by
rest of the rx path.

As there are no non-compact versions of the hal APIs, move remaining rx_ops
to hal_ops to use in non-performance critical path.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.3.1-00173-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Change-Id: I5b1688d80cceefda05174e33f0436c9646003e0c
Signed-off-by: Pavankumar Nandeshwar <quic_pnandesh@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>