mac80211: fix dynamic vlan warning with monitor interface restart
When monitor interface restarts, in nss offload disabled
case, the encap and decap offload flags are removed
from all the interfaces in that phy#.

However when dynamic VLAN and monitor interfaces are
created in the same phy#, these flags are not updated
correctly, due to which warning calltrace is observed.

Add condition check to update the correct flags in
dynamic VLAN case.

Change-Id: I3d5a9d55e9036ea247151b6d7257b41314543ac3
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>