mac80211: Mac Initialize EWMA fail avg to
In certain scenario with more number of pass frames,
ewma_fail_avg would reach 0(lowest value) and in ewma method
value with 0 is a tricky and that could give some higher value
even with single failure value and that triggers MPL broken logic.
Note that it is already initialized to 1 during the time of init
or new mesh path assignment.

Change-Id: I1b49582b67e7a36e3cb428243aa854c70a02e397
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>
Signed-off-by: Karthik M <quic_karm@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>