mac80211: athk Monitor mode bring up in SLO
Currently all the radios are sharing the same chandef, so during monitor
start wrong values are getting passed to FW, due to which we are seeing
FW assert.

Fixed by adding a check, when ar is not compatible with the chandef,
it will stop the process.

In addition, cfg80211 was not allowing to change the channels before
bringing up the interface, it needs at least one interface up to
change the channels.

Hence, removing the check for number of running interfaces.

Change-Id: I4da29931d49c9c3c46b89aba2df9f8f70ba5c561
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>