mac80211: Add support to enabledisable FILSUBPR frames
Add Support to enabled/disable the Fast initial Link Setup(FILS)
ands Unsolicited Broadcast Probe response frames in run time.

Currently FILS or UBPR config with interval 0 is considered
as invalid config, Allow user to set the fils max_interval and
ubpr interval to be set to 0 in which is case it is disabled.
This will allow run time disbable of FILS/UBPR rather doing
a BSS restart.

Change-Id: I3420d70f2374f7a1688a34b6a044cf5415339a9b
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Signed-off-by: Rameshkumar Sundaram <quic_ramess@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>