mac80211: pass channel info in AWGN event
Currently during AWGN event, drivers only passes the
ieee80211_hw on which AWGN interference is detected and
the interference bitmap. The worker, iterates over each
channel contexts avaiable under the hardware and expects
only 1 channel contexts to be present. Then, with this
channel context it executes further steps.

However, with MLO, there can be more than 1 channel contexts
available under the hardware hence while iterating, it is
important to know on which channel AWGN interference was
detected.

Add logic to pass channel info as well in the AWGN detcted
event and use it properly to process further in the AWGN
event.

Change-Id: I059b055c2cb2c4b24dd653098d38b7bf8a1e6d3b
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>