mac80211: RU puncturing support for STA
Currently, when AP detects an radar in dfs channel then channels are punctured
in which radar is detected but STA is downgrading its bandwidth instead of
handling this puncturing.

For example if AP is in EHT80 mode channel 100 and radar is detected, channel
puncturing takes place. Observed that STA reduces its bandwidth to 40 MHz. This
should not happen, STA should remain in 80Mhz bandwidth.

We are not handling puncturing bitmap on STA side. When AP advertises puncturing
bitmap, STA should handle this puncturing.

To resolve this issue, we should parse EHT operational IE and update chandef
with puncturing bitmap for EHT sta.

Change-Id: I75ab814537537522e5164151b268c4a5f1f218f0
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>