mac80211: fix scan issue due to single wiphy architecture
Currently, irrespective of the links active in an interface, if frequency
info is not provided in the scan request, all channels across all the
available bands under the wiphy are selected. This leads to huge scan request
as well as long time to complete the scan.

However, if an interface is active on a certain num of links, then there
is no point in scanning for all the bands and their channels.

Add fix to select channels only from bands which are currently active on
the requested scanning interface.

Channels are freed when the pdev bring up is failed during recovery due to
firmware crash. Need to do sanity check if the channel is freed for a
particular band.

Change-Id: Ia35b64fcdd2482733e3f659a984c4ed9ba18a590
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>