mac80211: wifi export DFS CAC usable state helper function
cfg80211 has cfg80211_chandef_dfs_usable() function to know whether at
least one channel in the chandef is in usable state or not.

Make the above function visible to drivers by exporting its symbol
to global list of kernel symbols.

Lower level drivers can make use of this function to be aware
if CAC is required on the given chandef. For example drivers which
maintains the CAC state internally can make use of this.

Change-Id: Id3c9e3487b15d0d2e86870d48a1e1d2d60e832e0
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>