mac80211: wifi Fix station dump fields in WDS mode
After WDS AP gets associated with WDS STA, DTIM and beacon interval
values in the station dump is seen as 0.

During the WDS Association, DTIM and beacon interval stats are being referred
from the VLAN interface instead of the AP's interface. Fix this issue by
referring to the AP's interface to get the DTIM and beacon interval values.

Change-Id: I22bded6bade29d74e6ff5fe0f73adb4bf06675a4
Signed-off-by: Ganesh Kariganuru Mahabalesh <quic_gkarigan@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>