mac80211: wifi add radar handling for device BW parameters
Use the device bandwidth parameters (center frequency and width) instead
of the operating bandwidth parameters to set the DFS state for channels
if the values are different. Include device bandwidth and center freqeuncy
parameters in the radar events sent to the userspace.
When radar is detected in the non operating portion of the device
bandwidth, handle it depending on the current state of the affected
channels:
- If the channels are available, mark those as unavailable and send
the radar detection event to userspace.
- If the channels are already unavailable, ignore the radar
event until the no occupancy period expires.
- If the channels are usable and if a background CAC is running then
abort the CAC and send event to the userspace.

Change-Id: I521894bafca9419c7562ec2d803ea5d9422be656
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>