mac80211: wifi Set eht operation
Currently the ext tag: eht oper's basic mcs nss value is set to the
max tx and rx nss values. This is an issue because if the station's
rx and tx nss are below the basic mcs nss of the eht oper in AP, the
association will not happend in EHT mode(as per draft doc at EOM).

Hence the basic mcs nss is now set to be 0x00000011 instead of the max
nss values, so that the association between station and AP can now
be done with the EHT mode.

w.r.t. the draft doc.: IEEE 802.11-23/0638r4 "An EHT STA shall not
attempt to join (MLME-JOIN.request primitive) a BSS unless it supports
(i.e., is able to both transmit and receive using) all of the
<EHT-MCS, NSS> tuples in the basic EHT-MCS and NSS set."

Change-Id: Ifae474fe85c7f8886f176c15f1d1b7404a185839
Signed-off-by: Naveen S <quic_naves@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>