mac80211: wifi Add NL attribute to denote ft_auth sta flag
A few drivers may need to know if STA is using FT auth to decide on
allowing set key.

Add new netlink attribute NL80211_STA_FLAG_FT_AUTH to get the ft_auth
sta flag. This flag is used to set need_ptk_4_way to false so that the
driver is notified about the FT auth and allows set key.

Change-Id: I3237aa4a31e5c73c06660c765969c4997bb873d1
Signed-off-by: Amith A <quic_amitajit@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>