mac80211: wifi Set ft_auth flag for FT roaming
A few drivers may need to know if STA is using FT auth to decide on
allowing set key.

In order to let such drivers to allow set key, add a flag 'ft_auth'
to ieee80211_sta and set it to true in case of FT roaming. This will
be used to set the need_ptk_4_way to false and thereby allowing the
driver to set key.

Change-Id: Ia3bf6359bca2a33131651c0103c1fa66542a51ec
Signed-off-by: Amith A <quic_amitajit@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>