mac80211: Handle vendor command for interface offload type
Add support to handle interface offload type configuration
through vendor command from userspace tools. Use the offload
type configured to allocate respective PPE VP and map it with the
netdev. This enables the PPE HW to identify the flows and handle
the frames in the respective offload datapath configured

Patch-depdendency: 955-QSDK-wifi-ath12k-Handle-vendor-command-for-interface-offl.patch

Change-Id: I00473dc461e6c5683db9018d518d5642b9aadcc0
Signed-off-by: Balamurugan Mahalingam <quic_bmahalin@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>