mac80211: QSDK wifi Fix ieee set active links function
Currently, in ieee80211_set_active_links does call drv_change_sta_links
and then call link_use_channel. The problem is that drv_change_sta_links
will update the resources to the firmware post recovery before the
creation or assignment of chanctx for the targeted vif which resulted
in unexpected behavior.

The caller must ensure that link are assigned with an existing or new
chanctx before changing the links resources in the target.

Change list:
        - Fix Mode-0 SSR for WDS STA mode - by calling link_use channel
        function before drv_change_sta_links.

Patch-dependency: None
Patch-work: None

Change-Id: Iade1a97c421899ee9c5462fd6d776a5253983e1d
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>