mac80211: wifi Handle link addresses separately in ieee80211_get_station()
Currently, ieee80211_get_station() functions correctly for an MLD
station only when the MLD MAC address is provided. This is because
the MLD station is stored based on its MLD address. However, there
are instances where it is necessary to retrieve the MLD station
information based on the link address. In such cases, this function
would return -ENOENT since it would not find a station entry with
the link address.

To resolve this, we extend ieee80211_get_station() to check for link
addresses within the MLD stations. If a match is found, the function
will return the MLD station information, ensuring proper handling of
both MLD and link addresses.

Patch-dependency: [None]
Patch-work: [None]

Change-Id: Ife4ad0528291b9de25d19bb07004cc14c9f6e837
Signed-off-by: Santhosh Ramesh <quic_santrame@quicinc.com>
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>