mac80211: UPSTREAMDPwifi Fix KW issue bssid NULL dereferencing
Currently, when retrieving the BSSID value using the function
ieee80211_get_bssid(), it may return NULL.
Later, in ieee80211_rx_h_sta_process(), this value is used to
check the equality of Ethernet addresses with the function
ether_addr_equal().
Since ether_addr_equal() expects two valid addresses, this could
lead to a segmentation fault or other undefined behavior.

Hence, to prevent this add a null check for the BSSID.

Change-Id: Ied5f7d6d3b6f42b7f50b6b8b3706721b99d376fb
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>