mac80211: QSDKCP wifi add option for vif allowed radios
This allows users to prevent a vif from affecting radios other than the
configured ones. This can be useful in cases where e.g. an AP is running
on one radio, and triggering a scan on another radio should not disturb it.

Changing the allowed radios list for a vif is supported, but only while
it is down.

While it is possible to achieve the same by always explicitly specifying
a frequency list for scan requests and ensuring that the wrong channel/band
is never accidentally set on an unrelated interface, this change makes
multi-radio wiphy setups a lot easier to deal with for CLI users.

By itself, this patch only enforces the radio mask for scanning requests
and remain-on-channel. Follow-up changes build on this to limit configured
frequencies.

patch-work: None
patch-dependency: None

Link: https://lore.kernel.org/all/eefcb218780f71a1549875d149f1196486762756.1728462320.git-series.nbd@nbd.name/

Change-Id: I22b84375d43b75dfe86fa3a2dc86c2294f502719
Signed-off-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>