mac80211: QSDKPROTCOL Reorder work flushes after unlocking wiphy and rtnl locks
Reorder the flush operations for `awgn_detected_work` and `cw_detected_work`
in `ieee80211_unregister_hw` to occur after unlocking the wiphy and rtnl locks.
This change ensures that the work flushes are performed without holding
these locks, potentially preventing deadlocks and improving the overall
stability of the unregister process.

Change-Id: I995feef77775a0e428766316fd5c8197991b4c0a
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>