mac80211: QSDKPROTOCOL wifi Fix station association issue in MHz during bandwidth change for GHz interface
When the AP changes bandwidth to 40 MHz and station try to reconfig
new bandwidth (40 MHz) based on the received beacon from AP via
ieee80211_config_bw(), the function assign bandwidth based on HT
operation, but for 6 GHz band HT and VHT operation is not supported,
due to this 6 GHz station not able to get correct bandwidth (40 MHz)
and associate with default bandwidth (20 MHz).

To address this issue, allow bandwidth configuration based on HE and
EHT operation for 6 GHz interface.

Patch-dependency: None
patch-work: None

Change-Id: I5545387144605a508f2d9007bff23b76ff811642
Signed-off-by: Aaradhana Sahu <quic_aarasahu@quicinc.com>
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>