wifi: ath12k: Hardware specific ieee80211_ops registration
With unified driver, the Tx functionality is split between common and
architecture-specific modules. Defining the ieee80211_ops Tx callback
in the common module introduces additional indirections, as the common
Tx function calls the architecture-specific Tx functions via ops. To
avoid this, the Tx callback should be defined in the architecture-specific
module. This patch establishes the base framework to register the
ieee80211_ops table according to the underlying hardware architecture.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.3.1-00173-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Change-Id: I9c67dacda3d75463d7eb3101772203e07944084f
Signed-off-by: Ripan Deuri <quic_rdeuri@quicinc.com>
Signed-off-by: Harsh Kumar Bijlani <quic_hbijlani@quicinc.com>
Signed-off-by: Aman Mehta <quic_amanmeht@quicinc.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>