QSDK: mac80211: Fix warn on if mesh metric is zero
During performence measurement, stats is disabled through boost
performence script. So, host is not receiving the metric values
from firmware and the value becomes zero during mpath finding.
This issue addressed through static mpath. But,in RvR far range
test, connection becomes unstable and getting this warnon
continuously. So, add a check to prevent this warnon if the
stats is disabled.

Patch-dependency:
	913-001-wifi-mac80211-avoid-tx-pending-queue-in-native-wifi-.patch

Change-Id: Icba93fb80e79177e5350d74a2a2e750943e18f95
Signed-off-by: Thiraviyam Mariyappan <tmariyap@qti.qualcomm.com>
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>