QSDK:PLATFORM: wifi: ath12k: Enable amss_dualmac.bin loading for QCN9274
Add changes to load amss_dualmac.bin when firmware-2.bin is not
present.

If firmware-2.bin fails to load, fw features will not be populated, so
any checks on fw features should be skipped in fallback scenarios. So
adding a check on fw api version to bypass any fw features related checks.

Change-Id: Ic0885b8081ec04fdf4839264c6f1a531e2abb5cd
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>