wifi: ath12k: n-link-mlo bridge vdev create and start
This patch handles the bridge vdev create, start, stop and
delete during interface creation and deletion from userspace(hostapd)
during wifi up and down, and during SSR mode cases. Currently the
N-Link MLO is supported only for AP mode. In future, it will be
extended to STA and WDS mode too.

The changes in the patch are as follows:
	1) Handle bridge vdevs create, start, stop and delete
	   operations during normal interface addition/deletion
	   from userspace(hostapd) that is during wifi up/down.

	2) Two bridge vdevs for the MLD will be added once the second
	   link of the MLD is added successfully. Currently for
	   4 QCN9274 devices, when all the four normal links are
	   configured in an MLD, we need a maximum of 2 bridge vdevs.
	   These two bridge vdevs are selected in a way that, for
	   any MLD combinations, MLO communication is possible.
	   In our case, the bridge vdevs will be created in the
	   pairs 5 GHz-lower, 6G and 5 GHz-upper, 2 GHz. When the
	   bridge vdev count in the first pair exceeds,
	   the bridge vdevs will be created in the second pair.
	   The combination in the pair is dependent on the probe order
	   of the device. Depending on the order it is stored in
	   "ath12k_hw" and from there the combination in the pair is
	   selected.

	3) Bridge vdevs for the MLD will be added only when the second
	   link of the MLD is added successfully.

	4) Bridge vdevs of the MLD will be removed only after, all the
	   normal links of the MLD is removed successfully.

	5) When a bridge vdev needs to be created in a radio, where
	   even a single normal vdev is not yet configured, then basic
	   channel configurations will be extracted from the radio
	   configurations and those will be used for the bridge vdev.

	6) For firmware backward compactability, these service flags(
	   WMI_TLV_SERVICE_N_LINK_MLO_SUPPORT,
	   WMI_TLV_SERVICE_BRIDGE_VDEV_SUPPORT) denotes whether the
	   firmware supports N-link MLO feature.

	7) Handle bridge vdevs reconfiguration during SSR mode scenarios.

	8) Handle link stats for bridge vdevs.

In STA mode, similar to AP mode, a bridge vdev is added when links
are established between diagonally opposite devices in an MLD.
However, in STA mode, a single bridge vdev suffices for MLO
communications between these diagonally opposite devices. Therefore,
only one bridge vdev will be created in the adjacent device of the
diagonally opposite devices.

Here the adjacent devices where the bridge is created are
determined using the “device_bitmap” stored in “ahvif”. This bitmap
represents the device_idx of the vdevs created under that “ahvif”.
With this information, the devices where a bridge vdev can be
created are selected based on the adjacent device information of the
currently created vdevs. If the first choice of the adjacent
device cannot create a bridge vdev due to max bridge vdevs allocated
or any other errors, it will move to the next adjacent device for
bridge vdev creation. And even the second adjacent device cannot
create bridge vdev, it will exit gracefully as in AP mode.

Patch-dependency:
	0327-01-wifi-ath12k-n-link-mlo-bridge-vdev-basic-handling.patch

Tag: N_LINK_MLO

Change-Id: Id4f0cf2bc9fa517671ecd5bb26985126dc227233
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>