ath12k: update all 6 GHZ reg rules in sband
Based on the power modes in 6 GHZ band, there can be total of 9
possible channel arrays in sband. This patch updates the 6G
sband with all possible 9 modes. The sband will be updated
with default values since we dont know the power mode during
sband setup (initial boot).

Once the FW sends the reg rules and regd is built, cfg80211
will update the sband as per the rules. To aid cfg80211 to
distinguish between the power mode, the ath11k updates the
power mode as well while building regulatory domain.

Change-Id: I8f0ac7069e6899a7a8e893cdd6f6c0b1dd328621
Signed-off-by: Aishwarya R <aisr@qti.qualcomm.com>
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>