ath12k: Add support to bring in debug levels
Currently there is no debug level logging mechanism present in the
driver, so when enabling debug mask all the debug logs are displayed.
Add support to bring in debug levels for every debug mask, which
only gives the log of specific level which requires and add
debug mask to view the log of mac delete id, type, subtype, map.

Working Mechanism:

By comparing the leftmost bit with the debug level(level0 = 0,
level1 = 1, level2 = 2), the logs of corresponding level are
shown, which enables specific logs to be shown while avoiding an
overflow of unwanted logs.

TODO: few hunks in mac_vdev_delete() and Critical update are yet
to be added.

Change-Id: I7a21b7d0c34ebf48149147a98f0dfc1f7f44e9e5
Signed-off-by: Gunaal R <gunar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>