QSDK:PLATFORM: wifi: ath12k: Add support to download rxgainlut.bin
Update the QMI sequence to download rxgainlut bin files, if rxgainlut
support is enabled in firmware, firmware will advertise in the
firmware capabalities based on that rxgainlut.bin is downloaded to
firmware over QMI.

Rxgainlut binaries are used as input parameters for front end modules
(FEM) configuration.

If it fails to download rxgainlut from board-2.bin, it will check for
rxgainlut.b<board_id> file in /lib/firmware/ directory, if it fails to
download, then it will check for rxgainlut.bin file.

Change-Id: I7f31364c33cec1808a48b5fc439dd0d97f9dad7d
Signed-off-by: Saleemuddin Shaik <saleemud@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>