UPSTREAM: DP: wifi: ath12k: Update pending desc size in reo cache flush
During rx_tid_del, if the desc size is greater than
128 bytes, total desc size will be splitted into
respetive 128 byte low and high address format and
updated in flush cache descriptor. During failure
case, we will have the entire rx_tid will be saved
to reo_cmd list to free in next cycle. In a scenario,
where one 128 byte desc is actually freed and the
next desc is failed, we are still saving the entrie
rx_tid and in the next cycle we will try to flush
the same desc again which results in memory corruption.

To fix this, update pending desc size to rx_tid struct
and use that in next cycle while calculating total
desc size in order to avoid updating same desc again.

Change-Id: I7a440bf6791ee411b1bd9758c73fdeff991c6b78
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>