UPSTREAM: DP: wifi: ath12k: Add Retry Mechanism for REO RX Queue Update Failures
Currently, under stress scenarios, when the REO command ring is
full, the peer delete update for the RX queue fails due to lack
of space to send the command. In such cases, the host performs
a dma_unmap and frees the allocated memory. However, the hardware
still retains a reference to the same virtual address. During the
next allocation cycle, the kernel may reassign this address,
unaware that the hardware is still using it. This can lead to
memory corruption, as the host may access or write to memory
that is still in use by the hardware.

To prevent memory corruption, a retry mechanism is added for
HAL_REO_CMD_UPDATE_RX_QUEUE. This involves introducing a new
list in the dp structure, protected by a dedicated lock to
ensure safe access. With this change, the host no longer frees
memory immediately on failure. Instead, the memory is freed
only after the hardware has released the corresponding virtual
address.

Additionally, the DP_REO_CMD_RING_SIZE is increased to 256 to
reduce the likelihood of ring exhaustion. Furthermore, checks
are added for rx_tid->vaddr during reo_cmd_list_cleanup() and
dp_reo_cmd_free() to avoid double free or unmap operations.

Change-Id: If74fed5fc2d7513a46a0504b6ff6e85e000e0f00
Signed-off-by: Nithyanantham Paramasivam <nithp@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>