wifi: ath12k: add logic to select primary link
Current logic of primary selection uses debugfs as the only way to select
primary link. If debugfs is not provided, it will go with the assoc link
as primary link. In order to make primary link selection logic more robust,
following changes are being made -

1. If debugfs entry is there, give preference to it.
2. If it is not there, then use preferable hw_param enabled links as
   candidate links.
3. If no preferable hw_param enabled link exist, then consider whatever
   links are present as candidate links.
4. Now among this candidate links, select first set bit.

TODO: Enhance logic to use RSSI or some other metric to decide the
      primary link.

Change-Id: I205a579aba1631239716d43c21e9b36bb0edeb7c
Signed-off-by: Aditya Kumar Singh <adisi@qti.qualcomm.com>
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>