wifi: ath12k: process UMAC migration user request
In order to support pirmary UMAC migration, add the base infrastucture
to process the incoming request and to form the list which can be later
used to send command to firmware.

Since in a single shot, multiple peers could be migrated, a linked list
is being maintained which is ready to be migrated. A subsequent change
will accept this list and send command to firmware.

Change-Id: I9f7ff4b7622981ca99c6df14dec57998cdee4c43
Signed-off-by: Aditya Kumar Singh <quic_adisi@qti.qualcomm.com>
Co-developed-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>