UPSTREAM:PROTOCOL:wifi:ath12k: Generate SP rules based on AFC payload
Currently, after an AFC payload is received, the frequency object of
the AFC payload is coalesced and stitched together to form continuous
SP rules. These rules are then updated in the wiphy regd. The PSD
power of the coalesced rule is set to the minimum PSD power of the
individual AFC payloads.

This has the disadvantage of truncating the SP rules present in the
AFC payload, resulting in a loss of information and sub-optimal
performance.

To fix this generate new SP rules from AFC payload for updating in
the wiphy regd.

After AFC payload is received and processed, update the kernel regd
such that the reg rules for the SP channels enabled by AFC payload
are marked as enabled. AFC payload processing is updated as follows:
    (1) For a Valid payload, coalesce the AFC frequency objects to form
        contiguous frequency ranges. Two adjacent frequencies (diff by
        1 MHz) is considered overlapped.
    (2) Ignore the AFC frequency ranges that are out of range the
        radio's HW frequency range.
    (3) Since cfg80211 checks for at least 20 MHz BW for a channel to
        be considered valid. Make sure that each coalesced frequency
        range is at least 20 MHz wide.
    (4) Also, validate that each 20 MHz channels in the frequency range
        is also present in the AFC payload 20 MHz opclass channel
        object.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch

Change-Id: I91dda9cc09a7dac236122f13cd4b863bc37434bf
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>