wifi: ath12k: add WMI infra support for PTQM
Previous change added support to form a list having peers ready for
migration.

Add the WMI infra to send the migration command to firmware. Multiple
peers could be packed in a single WMI command. The num peers that can
be packed in a single command depends on the WMI message size. The list
formed earlier will be the input to WMI send infra.

Also, add support to handle the migration event response from the firmware.

Co-developed-by: Aditya Kumar Singh <quic_adisi@qti.qualcomm.com>
Change-Id: I599f7b43a4e9a23c76f5617250012a081de02d35
Signed-off-by: Aditya Kumar Singh <quic_adisi@qti.qualcomm.com>
Signed-off-by: Karthikeyan Kathirvel <kathirve@qti.qualcomm.com>
Co-developed-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>