UPSTREAM:PROTOCOL:wifi:ath12k: Generate NO_IR rules based on AFC payload
Create No IR ranges for those frequencies which are present in
the SP regulatory rules but not present in the final AFC frequency
ranges. The final AFC range here is the coalesced, HW frequency
limited, 20 MHz boundary aligned AFC frequency ranges which
are also present in the AFC payload channel object.

This helps to differentiate the frequencies that are enabled in the
AFC payload and those that are not allowed by AFC.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch
ath12k/0380-a-UPSTREAM-PROTOCOL-wifi-ath12k-AFC-payload-cleanup-an.patch
ath12k/0380-b-UPSTREAM-PROTOCOL-wifi-ath12k-Generate-SP-rules-base.patch

Change-Id: Ief96b64e802cfb0a88c2f4ff05d1e881181dd380
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>